refactoring to make it possible to book in advance rooms which are already reserved (now).
